+++ app/src/main/java/ru/stoliarenkoas/weatherapp/model/WeatherWind.java
@@ -10,9 +10,9 @@ import lombok.ToString;
 public class WeatherWind {
     private static final String[] DIRECTIONS = {"East", "North-East", "North", "North-West", "West", "South-West", "South", "South-East"};
     private float speed;
-    @SerializedName("deg") private int degree;
+    @SerializedName("deg") private float degree;
 
     public String getDirection() {
-        return DIRECTIONS[((degree-22)/45)%8];
+        return DIRECTIONS[(((int)degree-22)/45)%8];
     }
 }
